+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+- Add more configuration values to the VS Code extension.
+
 ### Fixed
 - VS Code needed restart to apply kubectl config/context change. Closes [316](https://github.com/metalbear-co/mirrord/issues/316).
 - Fixed DNS feature causing crash on macOS on invalid DNS name due to mismatch of return codes. [#321](https://github.com/metalbear-co/mirrord/issues/321).